change mean dtype to torch.float in mamba.py l364

+++ mamba.py
@@ -361,6 +361,6 @@ class RMSNorm(nn.Module):
         self.weight = nn.Parameter(torch.ones(d_model))
 
     def forward(self, x):
-        output = x * torch.rsqrt(x.pow(2).mean(-1, keepdim=True) + self.eps) * self.weight
+        output = x * torch.rsqrt(x.pow(2).mean(-1, keepdim=True, dtype = torch.float) + self.eps) * self.weight
 
         return output
